viafb: suspend/resume support
This patch adds support for suspend/resume of the
VIA framebuffer device. It requires a version of OFW
that restores the video mode.

This patch is OLPC-specific as the proper upstream solution
is to move the VIA video path to using the kernel modesetting
infrastructure and doing a proper save/restore in the kernel.

viafb: Restore the Second Display Memory address (#10548)

This value is getting reset to 0x00 during a suspend/resume cycle
pointing the display back to the default framebuffer.  This is wrong
for displays rotated with a shadowfb.  Store and replace the memory
address for the Second display.  A complete suspend/resume register
restore should be implemented but for speed on the XO we try to
minimize the registers restored.

This fixes a problem where the software-rotated display was returned
to standard orientation after a S/R cycle.

viafb: ReInit 3d texture engine on resume (#10525)

This fixes a problem where the XV Texture adapter does not work
after a suspend resume cycle.

viafb: Add full suspend/resume for the VX855 chipsets

Now that userspace is using more functionality of the VX855
chipset we need to make sure that viafb knows how to re-init and
restore all the hardware pieces.  We break the via_suspend and
via_resume functions out into their own file and add the following
functionality.

Properly update the VGA HW regs, SEQ regs, and CRT regs.  We make
sure to exclude any regs that cause the hardware to act funny.

Save and Restore the HQV and Video Regs.  This includes a hack that
I added that uses a previously unused register to signal the userspace
code that it should force a video overlay update.

Re-initialize the 3d engine.  We are using the 3d engine for accelerating
composting for display rotation.  We have to make sure everthing is
re-initialized or suspend/resume doesn't work properly when rotated.

viafb: Don't restore certain Seq registers

There are some chipsets registers that we should not be restoring
on VT switch or suspend/resume.  These include things that are
setup by the bios or could be used by other drivers.

Also add restore of register 0x40. This was disabled for debugging only
and left forgotten til now.

Fixes bug where video gets stopped on resume.
http://mailman.laptop.org/pipermail/devel/2010-October/030220.html

viafb: Fix ARGB Cursor Suspend/Resume on XO 1.5

If Hardware Cursors were enabled on the XO 1.5 the cursor would
disappear on suspend/resume.  This adds the necessary registers
to restore the fboffset, position and mode on the 2nd display
channel used by the XO 1.5.  For this to work properly we also
need a patched chrome driver.  the ARGB Cursor position regs
are write only so I store the location values in the straight
HW Cursor position reg.  This is not 100% correct, but not necessarily
wrong.

viafb: Better state handling of HWCursor during S/R

We need to track the HWCursor mode so we don't enable
the cursor in suspend/resume when on a VT.  We also
need to restore the HWCursor position as well as the ARGB
since we are using that to track the position.