Input: enable i8042-level wakeup control
The OLPC XO laptop is able to use the PS/2 controller as a wakeup source.
When used as a wakeup source, the key press/mouse motion must not be lost.

Add infrastructure to allow controllers to be marked as wakeup-capable,
and avoid doing power control/reset on the i8042/serio/input devices when
running in this mode. For systems where this functionality is available,
you are expected to enable wakeups on the i8042 device, the serio
devices, and the relevant input devices, to ensure that the hardware is
left powered and untouched throughout the suspend/resume.

Signed-off-by: Daniel Drake <dsd@laptop.org>

NOTE: This commit has been modified from the upstream-submitted
patches floating around, to always enable wakeups. This is because
the final upstream design isn't finalised, so the required userspace
interface to enable this functionality could change uncomfortably.
Force-enable wakeups in the kernel to make things easier here, no
userspace interfacing needed.