rtc-cmos: report wakeups from interrupt handler
this change complements 5d0b7b9.  that commit ensured that after
a wakeup userspace can determine whether the RTC likely caused
that wakeup.  that only affects actual wakeups, however.  if the
RTC alarm expires before the system has slept, it needs to abort
the in-progress suspend.  for this the wakeup event needs to come
from the interrupt handler.

Signed-off-by: Paul Fox <pgf@laptop.org>