OLPC DCON: make con_cursor check blank state (#11940)
Having the DCON code call fb_blank (as it currently does) is probably
not quite correct.

In this case, the existing "am I alive" checks in fbcon_cursor all pass
when fbcon_cursor() is called by hide_cursor() (while printing a kernel
message) after the framebuffer has been blanked by the DCON.

Add in another check in the framebuffer layer to avoid this.
Going forward, the DCON code probably needs a bit of a rethink
in this area.

This fixes a case when moving from the activation GUI to the boot
animation, a cursor box was drawn in the bottom left, on top of the
boot animation. This was also seen to persist through to X.

Signed-off-by: Daniel Drake <dsd@laptop.org>