viafb: add command queueing mechanism (#12027)
We are seeing issues where the hardware completely hangs shortly after
resume when the corresponding suspend happened halfway through
a command sequence (related to the 3D engine).

To work around these hangs, hack the command queue processing into
the kernel, so that we can do it uninterrupted by suspend.

This isn't clean, but I think its the least invasive solution that we
can do for 12.1.0, which is just moving a chunk of code from userspace
to kernel space. For the future we will want to look at cleaner options
such as DRI/KMS drivers and/or openchrome.