via-camera: new horizontal range magic number for siv120d (#12623)
Even though siv120d and ov7670 seem happy to share the same camera
configuration on XO-1, XO-1.75 and XO-4, the XO-1.5 case has been
different. The ov7670-friendly settings cause what appears to be
bad horizontal sync in the captured image when siv120d is in use.

Experimentally, I determined that changing magic numbers between
the two sensor types solves this issue.

I also came up with an explanation for why there is a horizontal
line of green pixels at the bottom of the captured image, but I
was unable to fix it.

Signed-off-by: Daniel Drake <dsd@laptop.org>