libertas: fix loss of scan results
While processing scan results containing multiple BSS, the remainder of
the buffer is ignored if a probe response is found with a zero length
IE.  This causes "invalid IE fmt" if LBS_DEB_SCAN is enabled.

As a result, access points are omitted from scan results if there are
any OLPC XO-1 laptops nearby with mesh enabled, which is the default for
OLPC OS.  The mesh does not need to be in use.

As the number of XO-1 laptops increases, it becomes harder for the XO-1
to scan an access point.

As the number of access points increases, it becomes easier for the XO-1
to scan an access point.

This patch changes the scan to accept and ignore a zero length IE.

Fixes OLPC 12575.