Revert "xo_1.5_defconfig: add siv121c sensor"
This reverts commit b1cb46e589c1bed1e3b118bb6c7a545631455ac2.

The sensor requires a master clock of no more than 27 MHz, but the
XO-1.5 motherboard provides a master clock of 48 MHz.

The result is camera hangs and image corruption, which varies across the
sample space of tested units.

The XO-1.5 motherboard has a programmable clock generator but the clock
used by the camera is also tied to the USBCLK input of the VX855
companion chip, which uses it for USB host and HD audio.

Therefore the sensor is not supported on the XO-1.5 laptop.

Signed-off-by: James Cameron <quozl@laptop.org>